+++ src/model/mod.rs
@@ -1,4 +1,4 @@
-pub mod lua;
+pub mod dto;
 pub mod momotalk;
 
 use std::sync::Arc;